feat: support FieldMapper (WIP)
